move checking device credentials to AuthenticationService
